Move server building logic into build function in `startup.rs`
