+++ src/main.rs
@@ -1,10 +1,5 @@
-use std::net::TcpListener;
-
-use sqlx::postgres::PgPoolOptions;
-
 use email_newsletter::configuration::get_configuration;
-use email_newsletter::email_client::EmailClient;
-use email_newsletter::startup::run;
+use email_newsletter::startup::build;
 use email_newsletter::telemetry;
 
 #[tokio::main]
@@ -13,26 +8,8 @@ async fn main() -> Result<(), std::io::Error> {
     telemetry::init_subscriber(subscriber);
 
     let configuration = get_configuration().expect("Failed to read configuration.");
-    let connection_pool = PgPoolOptions::new()
-        .acquire_timeout(std::time::Duration::from_secs(5))
-        .connect_lazy_with(configuration.database.with_db());
 
-    let sender_email = configuration
-        .email_client
-        .sender()
-        .expect("Invalid sender address");
-    let timeout = configuration.email_client.timeout();
-    let email_client = EmailClient::new(
-        configuration.email_client.base_url,
-        sender_email,
-        configuration.email_client.authorization_token,
-        timeout,
-    );
-
-    let address = format!(
-        "{}:{}",
-        configuration.application.host, configuration.application.port
-    );
-    let listener = TcpListener::bind(address)?;
-    run(listener, connection_pool, email_client)?.await
+    let server = build(configuration).await?;
+    server.await?;
+    Ok(())
 }
